rename channel header menu component folder
