+++ webapp/src/components/channel_header_action/channel_header_menu.tsx
@@ -1,0 +1,41 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+import React from 'react';
+
+import JiraIcon from 'components/icon';
+
+export type Props = {
+    open: (channelId: string) => void;
+    channelId: string;
+    isInstanceInstalled: boolean;
+    userConnected: boolean;
+};
+
+export default class ChannelHeaderMenuAction extends React.PureComponent<Props> {
+    handleClick = (): void => {
+        const {open, channelId} = this.props;
+        open(channelId);
+    };
+
+    render(): React.ReactElement {
+        const {isInstanceInstalled, userConnected} = this.props;
+
+        if (!isInstanceInstalled || !userConnected) {
+            console.log(' IN HERE ');
+            return null;
+        }
+        return (
+            <React.Fragment>
+                <button
+                    className='style--none'
+                    role='presentation'
+                    onClick={this.handleClick}
+                >
+                    <JiraIcon type='menu'/>
+                    {'Jira Subscriptions'}
+                </button>
+            </React.Fragment>
+        );
+    }
+}
